I seperated the class and the rest of main
