+++ user.cpp
@@ -1,0 +1,10 @@
+#include "user.h"
+#include <string>
+#include <iostream>
+using namespace::std;
+
+User::User(string uname, string pword, int ID) 
+: username(uname), password(pword), id_num(ID)
+{}
+
+
