+++ user.h
@@ -1,0 +1,17 @@
+#include <iostream>
+#include <string>
+#include <vector>
+
+using namespace::std;
+
+class User 
+{
+ private:
+  string username;
+  string password;
+  int id_num;
+ public:
+  User(string, string, int);
+};
+
+
